requirements.txt
scipy==1.3.1
mxnet_cu101==1.5.1.post0
numpy==1.17.2
tqdm==4.31.1
opencv_python==4.0.0.21
Pillow==6.2.1
scikit_learn==0.21.3
PyYAML==5.1.2